Updated: refactored more h/w control from simctl.c to iosim.c
